+++ Source/AliveLibAE/Flash.cpp
@@ -4,13 +4,13 @@
 #include "Game.hpp"
 #include "stdlib.hpp"
 
-Flash::Flash(Layer layer, u8 r, u8 g, u8 b, s32 /*not_used*/, TPageAbr abr, s32 time)
+Flash::Flash(Layer layer, s32 r, s32 g, s32 b, s32 /*not_used*/, TPageAbr abr, s32 time)
     : EffectBase(layer, abr)
 {
     SetType(AETypes::eFlash_117);
-    field_6E_r = r;
-    field_70_g = g;
-    field_72_b = b;
+    field_6E_r = static_cast<s16>(r);
+    field_70_g = static_cast<s16>(g);
+    field_72_b = static_cast<s16>(b);
     field_78_flash_time = time + sGnFrame_5C1B84;
 }
