+++ Source/AliveLibAE/Flash.hpp
@@ -7,7 +7,7 @@
 class Flash final : public EffectBase
 {
 public:
-    Flash(Layer layer, u8 r, u8 g, u8 b, s32 /*not_used*/, TPageAbr abr, s32 time);
+    Flash(Layer layer, s32 r, s32 g, s32 b, s32 /*not_used*/, TPageAbr abr, s32 time);
 
     virtual void VUpdate() override;
 private:
