+++ Source/AliveLibAO/Flash.cpp
@@ -6,13 +6,13 @@
 
 namespace AO {
 
-Flash::Flash(Layer layer, u8 r, u8 g, u8 b, s32 /*not_used*/, TPageAbr abr, s32 time)
+Flash::Flash(Layer layer, s32 r, s32 g, s32 b, s32 /*not_used*/, TPageAbr abr, s32 time)
     : EffectBase(layer, abr)
 {
     field_4_typeId = Types::eFlash_81;
-    field_5E_r = r;
-    field_60_g = g;
-    field_62_b = b;
+    field_5E_r = static_cast<s16>(r);
+    field_60_g = static_cast<s16>(g);
+    field_62_b = static_cast<s16>(b);
     field_68_flash_time = time + gnFrameCount_507670;
 }
