+++ Source/AliveLibAO/ScreenWave.hpp
@@ -16,7 +16,7 @@ struct ScreenWave_Data;
 class ScreenWave final : public BaseGameObject
 {
 public:
-    ScreenWave(FP xpos, FP ypos, Layer layer, FP width, FP speed, s16 radius);
+    ScreenWave(FP xpos, FP ypos, Layer layer, FP width, FP speed, s32 radius);
     ~ScreenWave();
 
     virtual void VScreenChanged() override;
